+++ framework/views/_foreignTables.php
@@ -8,7 +8,7 @@
 if (!empty($foreignKeys)):?>
  * Has foreign keys to the tables:
  *
-<?php foreach ($foreignKeys as $relatedTable): ?>
- * - `<?= $relatedTable ?>`
+<?php foreach ($foreignKeys as $fkData): ?>
+ * - `<?= $fkData['relatedTable'] ?>`
 <?php endforeach;
 endif;