+++ framework/views/createTableMigration.php
@@ -16,7 +16,7 @@ use yii\db\Migration;
 /**
  * Handles the creation for table `<?= $table ?>`.
 <?= $this->render('_foreignTables', [
-    'foreignKeys' => $foreignKeys
+    'foreignKeys' => $foreignKeys,
 ]) ?>
  */
 class <?= $className ?> extends Migration