+++ tests/framework/console/controllers/MigrateControllerTestTrait.php
@@ -508,6 +508,144 @@ class {$class} extends Migration
 
 CODE;
         $this->assertEqualsWithoutLE($code, $file);
+
+        $class = 'm' . gmdate('ymd_His') . '_' . $migrationName;
+        $this->runMigrateControllerAction('create', [
+            $migrationName,
+            'useTablePrefix' => true,
+            'fields' => 'user_id:integer:foreignKey,
+                product_id:foreignKey:integer:unsigned:notNull,
+                order_id:integer:foreignKey(user_order):notNull,
+                created_at:dateTime:notNull',
+        ]);
+        $file = $this->parseNameClassMigration($class);
+        $code = <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles the creation for table `{{%test}}`.
+ * Has foreign keys to the tables:
+ *
+ * - `{{%user}}`
+ * - `{{%product}}`
+ * - `{{%user_order}}`
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->createTable('{{%test}}', [
+            'id' => \$this->primaryKey(),
+            'user_id' => \$this->integer(),
+            'product_id' => \$this->integer()->unsigned()->notNull(),
+            'order_id' => \$this->integer()->notNull(),
+            'created_at' => \$this->dateTime()->notNull(),
+        ]);
+
+        // creates index for column `user_id`
+        \$this->createIndex(
+            '{{%idx-test-user_id}}',
+            '{{%test}}',
+            'user_id'
+        );
+
+        // add foreign key for table `{{%user}}`
+        \$this->addForeignKey(
+            '{{%fk-test-user_id}}',
+            '{{%test}}',
+            'user_id',
+            '{{%user}}',
+            'id',
+            'CASCADE'
+        );
+
+        // creates index for column `product_id`
+        \$this->createIndex(
+            '{{%idx-test-product_id}}',
+            '{{%test}}',
+            'product_id'
+        );
+
+        // add foreign key for table `{{%product}}`
+        \$this->addForeignKey(
+            '{{%fk-test-product_id}}',
+            '{{%test}}',
+            'product_id',
+            '{{%product}}',
+            'id',
+            'CASCADE'
+        );
+
+        // creates index for column `order_id`
+        \$this->createIndex(
+            '{{%idx-test-order_id}}',
+            '{{%test}}',
+            'order_id'
+        );
+
+        // add foreign key for table `{{%user_order}}`
+        \$this->addForeignKey(
+            '{{%fk-test-order_id}}',
+            '{{%test}}',
+            'order_id',
+            '{{%user_order}}',
+            'id',
+            'CASCADE'
+        );
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        // drops foreign key for table `{{%user}}`
+        \$this->dropForeignKey(
+            '{{%fk-test-user_id}}',
+            '{{%test}}'
+        );
+
+        // drops index for column `user_id`
+        \$this->dropIndex(
+            '{{%idx-test-user_id}}',
+            '{{%test}}'
+        );
+
+        // drops foreign key for table `{{%product}}`
+        \$this->dropForeignKey(
+            '{{%fk-test-product_id}}',
+            '{{%test}}'
+        );
+
+        // drops index for column `product_id`
+        \$this->dropIndex(
+            '{{%idx-test-product_id}}',
+            '{{%test}}'
+        );
+
+        // drops foreign key for table `{{%user_order}}`
+        \$this->dropForeignKey(
+            '{{%fk-test-order_id}}',
+            '{{%test}}'
+        );
+
+        // drops index for column `order_id`
+        \$this->dropIndex(
+            '{{%idx-test-order_id}}',
+            '{{%test}}'
+        );
+
+        \$this->dropTable('{{%test}}');
+    }
+}
+
+CODE;
+        $this->assertEqualsWithoutLE($code, $file);
     }
 
     public function testGenerateDropMigration()
@@ -598,7 +736,10 @@ CODE;
         $class = 'm' . gmdate('ymd_His') . '_' . $migrationName;
         $this->runMigrateControllerAction('create', [
             $migrationName,
-            'fields' => 'title:string(10):notNull,body:text:notNull,price:money(11,2):notNull,created_at:dateTime'
+            'fields' => 'title:string(10):notNull,
+                body:text:notNull,
+                price:money(11,2):notNull,
+                created_at:dateTime'
         ]);
         $file = $this->parseNameClassMigration($class);
 
@@ -632,6 +773,285 @@ class {$class} extends Migration
         \$this->dropColumn('test', 'body');
         \$this->dropColumn('test', 'price');
         \$this->dropColumn('test', 'created_at');
+    }
+}
+
+CODE;
+        $this->assertEqualsWithoutLE($code, $file);
+
+        $migrationName = 'add_columns_to_test';
+        $class = 'm' . gmdate('ymd_His') . '_' . $migrationName;
+        $this->runMigrateControllerAction('create', [
+            $migrationName,
+            'fields' => 'user_id:integer:foreignKey,
+                product_id:foreignKey:integer:unsigned:notNull,
+                order_id:integer:foreignKey(user_order):notNull,
+                created_at:dateTime:notNull',
+        ]);
+        $file = $this->parseNameClassMigration($class);
+
+        $code = <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles adding columns to table `test`.
+ * Has foreign keys to the tables:
+ *
+ * - `user`
+ * - `product`
+ * - `user_order`
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->addColumn('test', 'user_id', \$this->integer());
+        \$this->addColumn('test', 'product_id', \$this->integer()->unsigned()->notNull());
+        \$this->addColumn('test', 'order_id', \$this->integer()->notNull());
+        \$this->addColumn('test', 'created_at', \$this->dateTime()->notNull());
+
+        // creates index for column `user_id`
+        \$this->createIndex(
+            'idx-test-user_id',
+            'test',
+            'user_id'
+        );
+
+        // add foreign key for table `user`
+        \$this->addForeignKey(
+            'fk-test-user_id',
+            'test',
+            'user_id',
+            'user',
+            'id',
+            'CASCADE'
+        );
+
+        // creates index for column `product_id`
+        \$this->createIndex(
+            'idx-test-product_id',
+            'test',
+            'product_id'
+        );
+
+        // add foreign key for table `product`
+        \$this->addForeignKey(
+            'fk-test-product_id',
+            'test',
+            'product_id',
+            'product',
+            'id',
+            'CASCADE'
+        );
+
+        // creates index for column `order_id`
+        \$this->createIndex(
+            'idx-test-order_id',
+            'test',
+            'order_id'
+        );
+
+        // add foreign key for table `user_order`
+        \$this->addForeignKey(
+            'fk-test-order_id',
+            'test',
+            'order_id',
+            'user_order',
+            'id',
+            'CASCADE'
+        );
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        // drops foreign key for table `user`
+        \$this->dropForeignKey(
+            'fk-test-user_id',
+            'test'
+        );
+
+        // drops index for column `user_id`
+        \$this->dropIndex(
+            'idx-test-user_id',
+            'test'
+        );
+
+        // drops foreign key for table `product`
+        \$this->dropForeignKey(
+            'fk-test-product_id',
+            'test'
+        );
+
+        // drops index for column `product_id`
+        \$this->dropIndex(
+            'idx-test-product_id',
+            'test'
+        );
+
+        // drops foreign key for table `user_order`
+        \$this->dropForeignKey(
+            'fk-test-order_id',
+            'test'
+        );
+
+        // drops index for column `order_id`
+        \$this->dropIndex(
+            'idx-test-order_id',
+            'test'
+        );
+
+        \$this->dropColumn('test', 'user_id');
+        \$this->dropColumn('test', 'product_id');
+        \$this->dropColumn('test', 'order_id');
+        \$this->dropColumn('test', 'created_at');
+    }
+}
+
+CODE;
+        $this->assertEqualsWithoutLE($code, $file);
+
+        $migrationName = 'add_columns_to_test';
+        $class = 'm' . gmdate('ymd_His') . '_' . $migrationName;
+        $this->runMigrateControllerAction('create', [
+            $migrationName,
+            'useTablePrefix' => true,
+            'fields' => 'user_id:integer:foreignKey,
+                product_id:foreignKey:integer:unsigned:notNull,
+                order_id:integer:foreignKey(user_order):notNull,
+                created_at:dateTime:notNull',
+        ]);
+        $file = $this->parseNameClassMigration($class);
+
+        $code = <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles adding columns to table `{{%test}}`.
+ * Has foreign keys to the tables:
+ *
+ * - `{{%user}}`
+ * - `{{%product}}`
+ * - `{{%user_order}}`
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->addColumn('{{%test}}', 'user_id', \$this->integer());
+        \$this->addColumn('{{%test}}', 'product_id', \$this->integer()->unsigned()->notNull());
+        \$this->addColumn('{{%test}}', 'order_id', \$this->integer()->notNull());
+        \$this->addColumn('{{%test}}', 'created_at', \$this->dateTime()->notNull());
+
+        // creates index for column `user_id`
+        \$this->createIndex(
+            '{{%idx-test-user_id}}',
+            '{{%test}}',
+            'user_id'
+        );
+
+        // add foreign key for table `{{%user}}`
+        \$this->addForeignKey(
+            '{{%fk-test-user_id}}',
+            '{{%test}}',
+            'user_id',
+            '{{%user}}',
+            'id',
+            'CASCADE'
+        );
+
+        // creates index for column `product_id`
+        \$this->createIndex(
+            '{{%idx-test-product_id}}',
+            '{{%test}}',
+            'product_id'
+        );
+
+        // add foreign key for table `{{%product}}`
+        \$this->addForeignKey(
+            '{{%fk-test-product_id}}',
+            '{{%test}}',
+            'product_id',
+            '{{%product}}',
+            'id',
+            'CASCADE'
+        );
+
+        // creates index for column `order_id`
+        \$this->createIndex(
+            '{{%idx-test-order_id}}',
+            '{{%test}}',
+            'order_id'
+        );
+
+        // add foreign key for table `{{%user_order}}`
+        \$this->addForeignKey(
+            '{{%fk-test-order_id}}',
+            '{{%test}}',
+            'order_id',
+            '{{%user_order}}',
+            'id',
+            'CASCADE'
+        );
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        // drops foreign key for table `{{%user}}`
+        \$this->dropForeignKey(
+            '{{%fk-test-user_id}}',
+            '{{%test}}'
+        );
+
+        // drops index for column `user_id`
+        \$this->dropIndex(
+            '{{%idx-test-user_id}}',
+            '{{%test}}'
+        );
+
+        // drops foreign key for table `{{%product}}`
+        \$this->dropForeignKey(
+            '{{%fk-test-product_id}}',
+            '{{%test}}'
+        );
+
+        // drops index for column `product_id`
+        \$this->dropIndex(
+            '{{%idx-test-product_id}}',
+            '{{%test}}'
+        );
+
+        // drops foreign key for table `{{%user_order}}`
+        \$this->dropForeignKey(
+            '{{%fk-test-order_id}}',
+            '{{%test}}'
+        );
+
+        // drops index for column `order_id`
+        \$this->dropIndex(
+            '{{%idx-test-order_id}}',
+            '{{%test}}'
+        );
+
+        \$this->dropColumn('{{%test}}', 'user_id');
+        \$this->dropColumn('{{%test}}', 'product_id');
+        \$this->dropColumn('{{%test}}', 'order_id');
+        \$this->dropColumn('{{%test}}', 'created_at');
     }
 }